new aas controller extension (#14)
