--- edc-controlplane/edc-controlplane-memory/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-net.catenax.edc.controlplane.ControlPlaneExtension
