+++ edc-extensions/aas-controller/README.md
@@ -1,0 +1,3 @@
+# Asset Administration Shell Controller
+
+This EDC extension implements a custom controller für the 'Asset Administration Shell'-Wrapper.
