+++ edc-extensions/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>net.catenax.edc</groupId>
+        <artifactId>catena-x-edc-parent</artifactId>
+        <version>0.0.2-SNAPSHOT</version>
+    </parent>
+
+    <groupId>net.catenax.edc.extensions</groupId>
+    <artifactId>edc-extensions</artifactId>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>aas-controller</module>
+    </modules>
+
+</project>
